revert LP64 audit changes: untested; patch saved in patches/lp64-audit-pending.patch
FossilOrigin-Name: a379ec3366a5cdb4c0179d78c9ac6c7813973365a18e138126b51b77d474af2c

+++ src/atk/extensions/metax.c
@@ -298,8 +298,8 @@ struct classheader *classID;
 	return FALSE;
     }
     
-    style_SetNewLeftMargin(columns, style_LeftMargin, (long)550000, style_CM);
-    style_SetNewIndentation(columns, style_LeftMargin, (long)-550000, style_CM);
+    style_SetNewLeftMargin(columns, style_LeftMargin, 550000, style_CM); 
+    style_SetNewIndentation(columns, style_LeftMargin, -550000, style_CM);
     style_SetFontSize(heading,style_ConstantFontSize,20);
     style_SetJustification(heading,style_Centered);
     style_SetFontFamily(fixed, "AndyType");
